cli/pyproject.toml
[project]
name = "aicoe-links"
version = "0.1.0"
description = "CLI admin tool for aicoe.fit link shortener"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
]

[project.scripts]
links = "links:cli"
